html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-location-assign.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: location.assign</title>
<iframe data-src="support/blank.htm?nav" src="support/blank.htm?src" loading="lazy"></iframe>
<script>
const iframe = document.querySelector('iframe');
iframe.contentWindow.location.assign(iframe.dataset.src);
</script>
<!-- Loading testharness.js here is intentional. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?nav", location.href).href);
    done();
  }, 250);
};
</script>
